Finished Personal Logo
Finished that assignment and I am pretty proud of it - I think that I started to fall more in love with it as i got further into the assignment and explored some other concepts. I think that of all of them I really like just my initials JS with the quill going through it, I think that the simplicity was really cool.